RDEN-11 🖋 Refactor in folder structure
React folder structure are implemented. (by this url https://www.taniarascia.com/react-architecture-directory-structure/)
