+++ client/src/components/Logo/LogoImage.js
@@ -1,0 +1,11 @@
+import logo from "assets/images/logo.png";
+
+// ==============================|| LOGO SVG ||============================== //
+
+const LogoImage = () => {
+  return (
+    <img src={logo} alt="Mantis" width="45" />
+  );
+};
+
+export default LogoImage;
