+++ client/src/layout/MainLayout/Header/HeaderContent/MobileSection.js
@@ -14,7 +14,7 @@ import {
 
 // project import
 import Search from "./Search";
-import Profile from "./Profile";
+import Profile from "./Profile/Profile";
 import Transitions from "components/@extended/Transitions";
 
 // assets
