+++ client/src/layout/MainLayout/MainLayout.js
@@ -7,8 +7,8 @@ import { useTheme } from '@mui/material/styles';
 import { Box, Toolbar, useMediaQuery } from '@mui/material';
 
 // project import
-import Drawer from './Drawer';
-import Header from './Header';
+import Drawer from './Drawer/Drawer';
+import Header from './Header/Header';
 import navigation from 'menu-items';
 import Breadcrumbs from 'components/@extended/Breadcrumbs';
