+++ client/src/layout/index.js
@@ -1,0 +1,3 @@
+import MainLayout from "./MainLayout/MainLayout";
+
+export { MainLayout };
