+++ client/src/routes/index.js
@@ -1,10 +1,10 @@
 import { useRoutes } from "react-router-dom";
 
 // project import
-import MainRoutes from "./MainRoutes";
+import MainRoute from "./MainRoute/MainRoute";
 
 // ==============================|| ROUTING RENDER ||============================== //
 
 export default function ThemeRoutes() {
-  return useRoutes([MainRoutes]);
+  return useRoutes([MainRoute]);
 }
