+++ client/src/services/index.js
@@ -1,0 +1,3 @@
+import { PatientService } from "./Patient.service";
+
+export { PatientService };
